// Presentation_Angular/ClientApp/src/app/Core/models/changeUserDetails.ts
export interface ChangeUserDetails {
  username: string;
  firstName: string;
  familyName: string;
  street: string;
  houseNr: string;
  houseBus: string;
  postalCode: string;
  city: string;
  streetBilling: string;
  houseNrBilling: string;
  houseBusBilling: string;
  postalCodeBilling: string;
  cityBilling: string;
}